+++ tests/JsonApi/Schema/Pagination/FixedPageBasedPaginationProviderTraitTest.php
@@ -7,8 +7,6 @@ namespace WoohooLabs\Yin\Tests\JsonApi\Schema\Pagination;
 use PHPUnit\Framework\TestCase;
 use WoohooLabs\Yin\Tests\JsonApi\Double\StubFixedPageBasedPaginationProvider;
 
-use function urldecode;
-
 class FixedPageBasedPaginationProviderTraitTest extends TestCase
 {
     /**
@@ -105,7 +103,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getSelfLink("https://example.com/api/users", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -118,7 +116,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getSelfLink("https://example.com/api/users?", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -131,7 +129,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getSelfLink("https://example.com/api/users?a=b", "a=c&b=d");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?a=c&b=d&page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?a=c&b=d&page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -144,7 +142,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getSelfLink("https://example.com/api/users", "page[number]=0&page[size]=0");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -157,7 +155,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getSelfLink("https://example.com/api/users?a=b", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?a=b&page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?a=b&page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -194,7 +192,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getFirstLink("https://example.com/api/users", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -231,7 +229,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getLastLink("https://example.com/api/users", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=5", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=5", $href);
     }
 
     /**
@@ -256,7 +254,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getPrevLink("https://example.com/api/users", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=4", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=4", $href);
     }
 
     /**
@@ -269,7 +267,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getPrevLink("https://example.com/api/users", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=1", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=1", $href);
     }
 
     /**
@@ -294,7 +292,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getNextLink("https://example.com/api/users", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=5", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=5", $href);
     }
 
     /**
@@ -307,7 +305,7 @@ class FixedPageBasedPaginationProviderTraitTest extends TestCase
         $link = $provider->getNextLink("https://example.com/api/users?", "");
         $href = $link !== null ? $link->getHref() : "";
 
-        $this->assertEquals("https://example.com/api/users?page[number]=3", urldecode($href));
+        $this->assertEquals("https://example.com/api/users?page%5Bnumber%5D=3", $href);
     }
 
     private function createProvider(int $totalItems, int $page, int $size): StubFixedPageBasedPaginationProvider